Compiling template with RTS.
For the class ExampleClassTemplateBase, I have defined three possible
template parameters: One, Two and Three, using the builtin types in C++.
These types are predefined in parameterTypes.H, which is included in the
make file ExmapleClassTemplateBases.C which is compiled. Inside the
ExampleClassTemplateBase.H I have defined the macros used to define the
typeName, as well as add the constructor to the RTS table for each
specialized classTemplate.